refactor: rename private module variable and clean up imports
- Rename `default_ingester_service` to `__DEFAULT_INGESTER_SERVICE` to indicate it's module-private
- Remove unused `IngesterService` class import from endpoints, scripts, and tests
- Fix test fixture naming and type hints in CLI tests
- Export `is_path_to_existing_file` from configuration module
- Clean up unused imports in config module
