+++ src/configuration/__init__.py
@@ -12,7 +12,7 @@ from .provider import (
     set_config_provider,
 )
 from .resolve import find_unresolved_directives, resolve_references
-from .utility import is_config_path
+from .utility import is_config_path, is_path_to_existing_file
 
 __all__ = [
     # config
@@ -21,6 +21,7 @@ __all__ = [
     "resolve_references",
     "find_unresolved_directives",
     "is_config_path",
+    "is_path_to_existing_file",
     # interface
     "ConfigLike",
     # provider
