+++ src/configuration/resolve.py
@@ -1,21 +1,14 @@
 from __future__ import annotations
 
-import contextlib
 import copy
-import io
 from abc import abstractmethod
-from datetime import datetime
-from inspect import isclass
-from os.path import join, normpath
 from pathlib import Path
-from typing import Any, Protocol, Type, runtime_checkable
+from typing import Any
 
 import pandas as pd
-import yaml
-from dotenv import load_dotenv
 from loguru import logger
 
-from src.utility import dget, dotexists, dotset, env2dict, replace_env_vars
+from src.utility import dget, dotexists, env2dict, replace_env_vars
 
 from .utility import is_config_path, is_path_to_existing_file, replace_references
